@@ -2,12 +2,12 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
 import { Logger } from '@app/core';
 import { ApiService } from '@app/services/api.service';
-import { NotifyService } from '@app/services/notify.service';
 import { TranslateService } from '@ngx-translate/core';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Group, DeviceAvailable, DevicesAvailable } from '@app/shared/models/group';
 import { HeaderService } from '@app/services/header-service';
 import { DatatableComponent } from '@swimlane/ngx-datatable';
+import { ToastrService } from 'ngx-toastr';
 
 const log = new Logger('GroupComponent');
 
@@ -17,8 +17,8 @@ const log = new Logger('GroupComponent');
   styleUrls: ['./group.component.scss']
 })
 export class GroupComponent implements OnInit {
-  @ViewChild('table') table: DatatableComponent;
-  @ViewChild('content') content: any;
+  @ViewChild('table', { static: false }) table: DatatableComponent;
+  @ViewChild('content', { static: false }) content: any;
   form: FormGroup;
   rows: Array<Group> = [];
   rowsTemp: any[] = [];
@@ -31,7 +31,7 @@ export class GroupComponent implements OnInit {
     private apiService: ApiService,
     private formBuilder: FormBuilder,
     private translate: TranslateService,
-    private notifyService: NotifyService,
+    private toastr: ToastrService,
     private headerService: HeaderService
   ) {}
 
@@ -116,7 +116,7 @@ export class GroupComponent implements OnInit {
       this.apiService.putZGroups(this.rows).subscribe(result => {
         log.debug(this.rows);
         this.hasEditing = false;
-        this.notifyService.notify();
+        this.toastr.success(this.translate.instant('api.global.succes.update.title'));
         this.apiService.getRestartNeeded().subscribe(restart => {
           if (restart.RestartNeeded) {
             this.headerService.setRestart(true);
@@ -150,7 +150,10 @@ export class GroupComponent implements OnInit {
   }
 
   open(content: any) {
-    this.modalService.open(content, { ariaLabelledBy: 'modal-basic-title' }).result.then(result => {}, reason => {});
+    this.modalService.open(content, { ariaLabelledBy: 'modal-basic-title' }).result.then(
+      result => {},
+      reason => {}
+    );
   }
 
   isFormValid(): boolean {
