@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
 import { Logger } from '@app/core';
-import { TranslateService } from '@ngx-translate/core';
-import { NotifyService } from '../services/notify.service';
+import { UnsubscribeOnDestroyAdapter } from '../shared/adapter/unsubscribe-adapter';
 
 const log = new Logger('AdminComponent');
 
@@ -10,8 +10,16 @@ const log = new Logger('AdminComponent');
   templateUrl: './admin.component.html',
   styleUrls: ['./admin.component.scss']
 })
-export class AdminComponent implements OnInit {
-  constructor(private notifyService: NotifyService, private translate: TranslateService) {}
+export class AdminComponent extends UnsubscribeOnDestroyAdapter implements OnInit {
+  action: string;
 
-  ngOnInit() {}
+  constructor(private activatedRoute: ActivatedRoute) {
+    super();
+  }
+
+  ngOnInit() {
+    this.subs.sink = this.activatedRoute.queryParamMap.subscribe(params => {
+      this.action = params.get('action');
+    });
+  }
 }
