@@ -7,6 +7,7 @@ import { SwResetComponent } from './sw-reset/sw-reset.component';
 import { ErasePdmComponent } from './erase-pdm/erase-pdm.component';
 import { ReloadPluginComponent } from './reload-plugin/reload-plugin.component';
 import { RescanGroupComponent } from './rescan-group/rescan-group.component';
+import { AssistProvisionningComponent } from './assist-provisionning/assist-provisionning.component';
 
 @NgModule({
   imports: [AdminRoutingModule, SharedModule],
@@ -16,7 +17,8 @@ import { RescanGroupComponent } from './rescan-group/rescan-group.component';
     SwResetComponent,
     ErasePdmComponent,
     ReloadPluginComponent,
-    RescanGroupComponent
+    RescanGroupComponent,
+    AssistProvisionningComponent
   ]
 })
 export class AdminModule {}
