@@ -1,24 +1,24 @@
 import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { CoreModule } from '@app/core';
 import { SharedModule } from '@app/shared';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { TranslateModule } from '@ngx-translate/core';
 import { NgBootstrapFormValidationModule } from 'ng-bootstrap-form-validation';
+import { ToastrModule } from 'ngx-toastr';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { DashboardModule } from './dashboard/dashboard.module';
-import { ShellModule } from './shell/shell.module';
-import { PushNotificationsModule } from 'ng-push';
 import { DeviceByNameComponent } from './dashboard/device-by-name/device-by-name.component';
+import { ShellModule } from './shell/shell.module';
 
 @NgModule({
   imports: [
     BrowserModule,
+    BrowserAnimationsModule,
     //    ServiceWorkerModule.register('./ngsw-worker.js', { enabled: environment.production }),
-    FormsModule,
     HttpClientModule,
     TranslateModule.forRoot(),
     NgBootstrapFormValidationModule.forRoot(),
@@ -27,7 +27,11 @@ import { DeviceByNameComponent } from './dashboard/device-by-name/device-by-name
     SharedModule,
     ShellModule,
     DashboardModule,
-    PushNotificationsModule,
+    ToastrModule.forRoot({
+      timeOut: 5000,
+      positionClass: 'toast-bottom-right',
+      preventDuplicates: true
+    }),
     AppRoutingModule
   ],
   declarations: [AppComponent],
