@@ -1,14 +1,23 @@
 import { Location } from '@angular/common';
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Observable, of } from 'rxjs';
-import { catchError, map } from 'rxjs/operators';
-import { PluginStats } from '@app/shared/models/plugin-stats';
-import { DomoticzEnv } from '@app/shared/models/domoticz-env';
-import { Plugin } from '@app/shared/models/plugin';
+import { Logger } from '@app/core';
+import { Capabilities } from '@app/shared/models/capabilities';
+import { Command } from '@app/shared/models/command';
 import { Device } from '@app/shared/models/device';
-import { Setting } from '@app/shared/models/setting';
+import { DomoticzEnv } from '@app/shared/models/domoticz-env';
 import { DevicesAvailable } from '@app/shared/models/group';
+import { NewDevice } from '@app/shared/models/new-hardware';
+import { Plugin } from '@app/shared/models/plugin';
+import { PluginStats } from '@app/shared/models/plugin-stats';
+import { Setting } from '@app/shared/models/setting';
+import { Observable, throwError } from 'rxjs';
+import { catchError, map } from 'rxjs/operators';
+import { DeviceByName } from '@app/shared/models/device-by-name';
+import { Binding } from '../shared/models/binding';
+import { ToastrService } from 'ngx-toastr';
+import { TranslateService } from '@ngx-translate/core';
+import { Relation } from '../shared/models/relation';
 
 const routes = {
   devices: '/device',
@@ -32,213 +41,245 @@ const routes = {
   domoticzEnv: '/domoticz-env',
   pluginHealth: '/plugin-health',
   rescanGroup: '/rescan-groups',
-  reqNwkfull: '/req-nwk-full'
+  reqNwkfull: '/req-nwk-full',
+  pluginRestart: '/plugin-restart',
+  devCommand: '/dev-command',
+  devCap: '/dev-cap',
+  newHardware: '/new-hrdwr/',
+  receiveNewHardware: '/rcv-nw-hrdwr',
+  binding: '/binding',
+  rawCommand: '/raw-command'
 };
+
+const log = new Logger('ApiService');
 
 @Injectable({ providedIn: 'root' })
 export class ApiService {
-  constructor(private httpClient: HttpClient, private location: Location) {}
+  constructor(
+    private httpClient: HttpClient,
+    private location: Location,
+    private toastr: ToastrService,
+    private translate: TranslateService
+  ) {}
 
   getPluginhealth(): Observable<Array<any>> {
     return this.httpClient.get(routes.pluginHealth).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getDevices(): Observable<Array<Device>> {
     return this.httpClient.get(routes.devices).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getZDevices(): Observable<any> {
     return this.httpClient.get(routes.zDevices).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getRawZDevices(): Observable<any> {
     return this.httpClient.get(routes.zdeviceRaw).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getZGroups(): Observable<any> {
     return this.httpClient.get(routes.zGroups).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getSettings(): Observable<any> {
     return this.httpClient.get(routes.settings).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   putSettings(settings: Array<Setting>): Observable<any> {
     return this.httpClient.put(routes.settings, settings).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getPlugin(): Observable<Plugin> {
     return this.httpClient.get(routes.plugin).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getPluginStats(): Observable<PluginStats> {
     return this.httpClient.get(routes.pluginStat).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getNwkStats(): Observable<Array<string>> {
     return this.httpClient.get(routes.nwkStat).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getTopologie(): Observable<Array<string>> {
     return this.httpClient.get(routes.topologie).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getReqTopologie(): Observable<any> {
     return this.httpClient.get(routes.reqTopology).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getReqInter(): Observable<any> {
     return this.httpClient.get(routes.reqInter).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getNwkFull(): Observable<any> {
     return this.httpClient.get(routes.reqNwkfull).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
-  getTopologieByTimeStamp(timestamp: string): Observable<any> {
+  getTopologieByTimeStamp(timestamp: string): Observable<Relation[]> {
     return this.httpClient.get(routes.topologie + '/' + timestamp).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getNwkStatsByTimeStamp(timestamp: string): Observable<any> {
     return this.httpClient.get(routes.nwkStat + '/' + timestamp).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   deleteTopologieByTimeStamp(timestamp: string): Observable<any> {
     return this.httpClient.delete(routes.topologie + '/' + timestamp).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   deleteNwkStatsByTimeStamp(timestamp: string): Observable<any> {
     return this.httpClient.delete(routes.nwkStat + '/' + timestamp).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getPermitToJoin(): Observable<any> {
     return this.httpClient.get(routes.permitToJoin).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   putPermitToJoin(settings: any): Observable<any> {
     return this.httpClient.put(routes.permitToJoin, settings).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
-  getZDeviceName(): Observable<PluginStats> {
+  getZDeviceName(): Observable<DeviceByName[]> {
     return this.httpClient.get(routes.zdeviceName).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   putZDeviceName(devices: any): Observable<PluginStats> {
     return this.httpClient.put(routes.zdeviceName, devices).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
+    );
+  }
+
+  deleteZDeviceName(_NwkId: string): Observable<PluginStats> {
+    return this.httpClient.delete(routes.zdeviceName + '/' + _NwkId).pipe(
+      map((body: any) => body),
+      catchError(error => this.handleError(error))
     );
   }
 
   getZGroupDevicesAvalaible(): Observable<Array<DevicesAvailable>> {
     return this.httpClient.get(routes.zgroupDevicesAvalaible).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   putZGroups(groups: any): Observable<any> {
     return this.httpClient.put(routes.zGroups, groups).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getZigate(): Observable<any> {
     return this.httpClient.get(routes.zigate).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getSwResetZigate(): Observable<any> {
     return this.httpClient.get(routes.swResetZigate).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getRescanGroup(): Observable<any> {
     return this.httpClient.get(routes.rescanGroup).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getZigateErasePDM(): Observable<any> {
     return this.httpClient.get(routes.zigateErasePDM).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
-  getReloadPlugin(plugin: Plugin, domoticzEnv: DomoticzEnv): Observable<any> {
+  getReloadPlugin() {
+    return this.httpClient.get(routes.pluginRestart).pipe(
+      map((body: any) => body),
+      catchError(error => this.handleError(error))
+    );
+  }
+
+  getReloadPluginOld(plugin: Plugin, domoticzEnv: DomoticzEnv): Observable<any> {
     const route =
       domoticzEnv.proto +
       '://' +
+      domoticzEnv.WebUserName +
+      ':' +
+      domoticzEnv.WebPassword +
+      '@' +
       domoticzEnv.host +
       ':' +
       domoticzEnv.port +
@@ -272,26 +313,85 @@ export class ApiService {
       plugin.Key +
       '&enabled=true&datatimeout=0';
 
+    const httpOptions = {
+      headers: new HttpHeaders({
+        'Access-Control-Allow-Origin': '*',
+        'Content-Type': 'text/plain'
+      })
+    };
+
     return this.httpClient
       .disableApiPrefix()
-      .get(route)
+      .get(route, httpOptions)
       .pipe(
         map((body: any) => body),
-        catchError(() => of('Error, could not load json from api'))
+        catchError(error => this.handleError(error))
       );
   }
 
   getRestartNeeded(): Observable<any> {
     return this.httpClient.get(routes.restartNeeded).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
   }
 
   getDomoticzEnv(): Observable<DomoticzEnv> {
     return this.httpClient.get(routes.domoticzEnv).pipe(
       map((body: any) => body),
-      catchError(() => of('Error, could not load json from api'))
+      catchError(error => this.handleError(error))
     );
+  }
+
+  getDevCap(id: string): Observable<Capabilities> {
+    const route = routes.devCap + '/' + id;
+    return this.httpClient.get(route).pipe(
+      map((body: any) => body),
+      catchError(error => this.handleError(error))
+    );
+  }
+
+  putDevCommand(command: Command): Observable<any> {
+    return this.httpClient.put(routes.devCommand, command).pipe(
+      map((body: any) => body),
+      catchError(error => this.handleError(error))
+    );
+  }
+
+  getNewHardware(enable: boolean): Observable<any> {
+    const enabled = enable ? 'enable' : 'disable';
+    const route = routes.newHardware + enabled;
+
+    return this.httpClient.get(route).pipe(
+      map((body: any) => body),
+      catchError(error => this.handleError(error))
+    );
+  }
+
+  getReceiveNewHardware(): Observable<NewDevice> {
+    return this.httpClient.get(routes.receiveNewHardware).pipe(
+      map((body: any) => body),
+      catchError(error => this.handleError(error))
+    );
+  }
+
+  putBinding(binding: Binding): Observable<any> {
+    return this.httpClient.put(routes.binding, binding).pipe(
+      map((body: any) => body),
+      catchError(error => this.handleError(error))
+    );
+  }
+
+  putCommandRaw(command: Command): Observable<any> {
+    return this.httpClient.put(routes.rawCommand, command).pipe(
+      map((body: any) => body),
+      catchError(error => this.handleError(error))
+    );
+  }
+
+  private handleError(error: any) {
+    log.error(error);
+    this.toastr.error(error.status + ' ' + error.statusText, this.translate.instant('api.global.error.title'));
+    return throwError(error);
   }
 }
