@@ -1,9 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-
 import { I18nService } from '@app/core';
-import { ApiService } from '../../services/api.service';
-import { NotifyService } from '@app/services/notify.service';
 import { HeaderService } from '@app/services/header-service';
+import { ToastrService } from 'ngx-toastr';
+import { ApiService } from '../../services/api.service';
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-header',
@@ -18,9 +18,10 @@ export class HeaderComponent implements OnInit {
 
   constructor(
     private headerService: HeaderService,
-    private notifyService: NotifyService,
+    private toastr: ToastrService,
     private i18nService: I18nService,
-    private apiService: ApiService
+    private apiService: ApiService,
+    private translate: TranslateService
   ) {}
 
   ngOnInit() {
@@ -40,12 +41,12 @@ export class HeaderComponent implements OnInit {
     if (event.currentTarget.checked) {
       this.permitToJoin.PermitToJoin = 255;
       this.apiService.putPermitToJoin(this.permitToJoin).subscribe((result: any) => {
-        this.notifyService.notify();
+        this.toastr.success(this.translate.instant('api.global.succes.update.title'));
       });
     } else {
       this.permitToJoin.PermitToJoin = 0;
       this.apiService.putPermitToJoin(this.permitToJoin).subscribe((result: any) => {
-        this.notifyService.notify();
+        this.toastr.success(this.translate.instant('api.global.succes.update.title'));
       });
     }
   }
