@@ -1,11 +1,14 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { extract } from '@app/core';
+import { BindingComponent } from './binding/binding.component';
+import { CommandComponent } from './command/command.component';
 import { ToolsComponent } from './tools.component';
 
 const routes: Routes = [
-  // Module is lazy loaded, see app-routing.module.ts
-  { path: '', component: ToolsComponent, data: { title: extract('tools') } }
+  { path: 'command', component: CommandComponent, data: { title: extract('command') } },
+  { path: 'binding', component: BindingComponent, data: { title: extract('binding') } },
+  { path: 'link', component: ToolsComponent, data: { title: extract('tools') } }
 ];
 
 @NgModule({
